Fix to work with tmux CC
With tmux 3.3a, with iTerm 3.5.0beta25, on Debian bookworm, this
script would fail to identify when it was running under tmux client
control (CC) mode, the mode which offers native OS windows on iTerm2.

This failure was because the TMUX env variable does not always have
the PID of the tmux process invoked with the -CC argument -- for
instance, in the case where there are multiple tmux sessions
running. The result of this failure was that pbcopy would think it was
not running under client control, use the wrong escape code, and print
a backslash into the terminal when reading input.

This commit fixes this, by relying on a different mechanism to
detect when pbcopy is running under tmux in client control mode.
The new mechanism is to ask tmux directly by checking the format
variable client_control_mode.